Add wrapper around WoW installs.
In the future this will provide a place to check if addons are
disabled and maybe poke other options. It also fixes passing
the install path around everywhere.

Also add a config version to migrate changes - v1 to v2 chops
the last two dirs off the install directory (to go from
_retail_/Interface/Addons to just _retail_).