Addon update rework
Allow individual IDs to be specified in the update command, to
only attempt updating those addons. (Default - no arguments -
is to update them all).

Change the update logic to prefer addons that are for the most
recent stable versions of the game (for now hardcoded to 8.x
and 7.x). This allows for the case where addons have betas
exclusively for Shadowlands (which will be ignored in favour
of older versions tagged for BfA) and addons which have only
9.x versions (which will be accepted).