Add a type field to addons in the config.
To unserialise an addon we now unserialise the BaseAddon
struct (which holds the type) then use that to construct
the correct instance.

For now everything is a CurseForgeAddon as the
add/update/etc code expects those fields.

Also add a migrate command that just forces a config
load/save. Not required for normal use but handy for
development.